feat(diaryProfiles): add enum for activity level
This enum will be used for DiaryProfileService.createWithCalculatedGoal to calculate goals

+++ src/main/java/com/github/plantfern/foodDiary/diaryProfiles/api/ActivityLevel.java
@@ -5,16 +5,16 @@ import lombok.Getter;
 
 
 public enum ActivityLevel {
-    SEDENTARY(1.2),
-    LIGHT(1.375),
-    MODERATE(1.55),
-    HIGH(1.725),
-    VERY_HIGH(1.9);
+    SEDENTARY(1.2F),
+    LIGHT(1.375F),
+    MODERATE(1.55F),
+    HIGH(1.725F),
+    VERY_HIGH(1.9F);
 
     @Getter
-    private final double coefficient;
+    private final float coefficient;
 
-    ActivityLevel(double coefficient){
+    ActivityLevel(float coefficient){
         this.coefficient = coefficient;
     }
 }
